Move data related modules to common submodule
